Fix Next route exports by moving auth helpers out of route file
